feat(esp-idf): Support for building with ESP-IDF

+++ lib/misc_mosq.c
@@ -225,11 +225,11 @@ char *misc__trimblanks(char *str)
 
 	if(str == NULL) return NULL;
 
-	while(isspace(str[0])){
+	while(isspace((int)str[0])){
 		str++;
 	}
 	endptr = &str[strlen(str)-1];
-	while(endptr > str && isspace(endptr[0])){
+	while(endptr > str && isspace((int)endptr[0])){
 		endptr[0] = '\0';
 		endptr--;
 	}
